Instrument content negotiation for relation convention
